chore: Update tsconfig.json and package.json
The `tsconfig.json` file was modified to change the `strict` flag from `true` to `false`. This change was made to relax the strict type checking in the TypeScript compiler.

The `package.json` file was modified to add new dependencies: `@vueuse/core`, `oh-vue-icons`, `vitest`, and `vue-router`. Additionally, the dev dependencies `@iconify-json/radix-icons` and `@iconify/vue` were added.

These changes were made to update the project dependencies and improve the development environment.