chore: Add cart functionality to Header.vue and ProductView.vue
The `Header.vue` component was modified to add cart functionality. A `CartOverview` component was imported and rendered conditionally based on the `cartVisible` flag. The cart icon was also updated to show the total quantity of items in the cart.

The `ProductView.vue` component was modified to add an "Add to cart" button. Clicking the button triggers the `addToCart` method from the `cartStore` and adds the selected product to the cart.

These changes were made to enable users to view and interact with the cart functionality in the application.